feat: Добавить трек в плейлист, Отобразить список плейлистов, Создать плейлист

+++ app/src/main/java/com/practicum/playlistmaker/media_library/data/FavoriteTrackRepositoryImpl.kt
@@ -27,7 +27,7 @@ class FavoriteTrackRepositoryImpl(
         return (tracks.map { track -> track.map { mapping.map(it) } })
     }
 
-    override suspend fun getFavoriteIdList(): List<Int> {
+    override suspend fun getFavoriteIdList(): List<Long> {
         return appDatabase.favoriteTracksDao().getFavoriteIdList()
     }
 }
