+++ app/src/main/java/com/practicum/playlistmaker/media_library/domain/api/FavoriteTrackRepository.kt
@@ -11,6 +11,6 @@ interface FavoriteTrackRepository {
 
     suspend fun getFavoriteTracks(): Flow<List<Track>>
 
-    suspend fun getFavoriteIdList(): List<Int>
+    suspend fun getFavoriteIdList(): List<Long>
 
 }
